Fix MSVC/GCC/Clang warnings in BRepGraph source and tests
- Remove redundant aStorage declaration hiding outer scope (EditorView)
- Rename inner aRefs to aRefsView to avoid shadowing (Tool)
- Replace break-loop with .More() for unreachable code fix (Validate)
- Initialize aAddStatus to avoid uninitialized warning (ShapesView)
- Change Count* helpers to return uint32_t (RefTestTools)
- Add static_cast for Size() in nbUniquePCurveNodes (Deduplicate test)

+++ src/ModelingData/TKBRep/GTests/BRepGraph_Deduplicate_Test.cxx
@@ -267,7 +267,7 @@ uint32_t nbUniquePCurveNodes(const BRepGraph& theGraph)
     const BRepGraph_EdgeId                              anEdgeId = anEdgeIt.CurrentId();
     const NCollection_LinearVector<BRepGraph_CoEdgeId>& aCoEdgeIdxs =
       theGraph.Topo().Edges().CoEdges(anEdgeId);
-    aCount += aCoEdgeIdxs.Size();
+    aCount += static_cast<uint32_t>(aCoEdgeIdxs.Size());
   }
   return aCount;
 }
